+++ src/datasources/data-frame/DataFrameDataSourceBase.test.ts
@@ -483,4 +483,20 @@ describe('DataFrameDataSourceBase', () => {
             expect(() => ds.transformResultQuery(filter)).toThrow('replace failed');
         });
     });
+
+    describe('parseColumnIdentifier', () => {
+        it('should return the empty column name and data type by default', () => {
+            const ds = new TestDataFrameDataSource(
+                instanceSettings,
+                backendSrv,
+                templateSrv
+            );
+
+            const parseResult = ds.parseColumnIdentifier('any-input');
+            expect(parseResult).toEqual({
+                columnName: '',
+                transformedDataType: ''
+            });
+        });
+    });
 });