+++ src/datasources/data-frame/DataFrameDataSourceBase.ts
@@ -119,6 +119,15 @@ export abstract class DataFrameDataSourceBase<
         return this.templateSrv.replace(filter);
     }
 
+    public parseColumnIdentifier(
+        _columnIdentifier: string
+    ): { columnName: string, transformedDataType: string } {
+        return {
+            columnName: '',
+            transformedDataType: ''
+        };
+    }
+
     protected constructNullFilters(columns: Column[]): ColumnFilter[] {
         return columns.flatMap(({ name, columnType, dataType }) => {
             const filters: ColumnFilter[] = [];